add landowner profile and profile details screen design, get user function

+++ lib/app/data/models/account.dart
@@ -32,7 +32,7 @@ class Account {
   String? email;
 
   @JsonKey(name: 'balance', defaultValue: 0)
-  int? balance;
+  double? balance;
 
   @JsonKey(name: 'point', defaultValue: 0)
   int? point;
